# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.CreateTraceRequest:
    properties:
      span_id:
        example: def456
        type: string
      tags:
        additionalProperties:
          type: string
        example:
          service: api
          version: "1.0"
        type: object
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
      trace_id:
        example: abc123
        type: string
    required:
    - span_id
    - trace_id
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      message:
        example: Field validation failed
        type: string
    type: object
...
  internal_handlers.TraceResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: "1"
        type: string
      message:
        example: Successfully processed
        type: string
      service_name:
        example: my-service
        type: string
      status:
        example: accepted
        type: string
      trace_id:
        example: abc123
        type: string
    type: object
...
host: '{{.Host}}'
info:
  contact: {}
  description: Distributed systems event replay backend
  title: Playback Backend API
  version: '{{.Version}}'
paths:
...
  /api/v1/traces:
    post:
      consumes:
      - application/json
      description: |-
        Create a new distributed trace
        Receive trace data from OpenTelemetry with optimized processing
      parameters:
      - description: Trace data
        in: body
        name: traces
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateTraceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.TraceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Receive traces
      tags:
      - traces
  /api/v1/traces/{id}:
    get:
      description: Get a trace by its ID
      parameters:
      - description: Trace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.TraceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get trace
      tags:
      - traces
...
swagger: "2.0"